Improve application healthcheck
Since we proxy our application through Nginx we need to be a bit more
precise than just checking that we can establish a connection on the
relevant port. We now verify that we're getting a response that doesn't
include:

  500 Internal Server Error
  502 Bad Gateway
  503 Service Unavailable

If we get a response code that doesn't match one of these values then we
assume that the application is running. We can't check for a 200 OK
because we don't know what routes are exposed on the host application
and what authentication is in place on them so e.g. a 404, 403, etc.
should be considered a valid response. It may end up being a better
choice to whitelist rather than blacklist - time will tell !