Simplify port maps and enable use outside of Rails
Local ports are now read from the docker-compose.yml rather than using
[pointless] magic to traverse config files. The docker-compose.yml is
the authority. We still use some config file munging to decide which
port to use at docker-compose.yml generation time.

Remove dependency on Rails for healthchecks. If we have a
docker-compose.yml then we have enough information to verify that a
service is running so we should be able to use the Rake tasks directly
from any Rails or non-Rails project.

Add a Rakefile which we can load into another project's Rakefile by
adding:

  load Orchestration.rakefile