Do not override host or port; use defaults instead
Also attempt to join Docker network that dependencies are running in.
This is a no-op if running on metal. If running in Docker (e.g. in
Jenkins where Docker socket is typically mounted rather than running
Docker-in-Docker) then containers will be sidecar-mounted so we need to
join their network to access them over DNS.