Fix volume configuration
Ensure named volumes are created for development and production mongo
and database containers. In test env, we use anon volumes.

Namespace volumes by environment to avoid collisions.

Although we expect users to remove the production volumes for
dependencies (i.e. if an app is deployed to a Docker Swarm then a
production database is unlikely to be running as a sidecar container but
rather as an entirely external dependency), there is a real-world
use-case for running single-node applications in production with a
single database. If this functionality isn't desired then the user can
simply remove these sections from the relevant compose files.