Simplify configuration detection and creation
Re-work configuration detection so that Orchestration dictate which
ports etc. our dependencies run on. We generate a relevant
docker-compose.yml files for each environment and set a random local
port for each. We then set this port in the relevant config files (e.g.
database.yml, mongoid.yml) and overwrite (with backups) any existing
config files.

This is a bit heavy-handed but it ensures we start with a clean slate
with consistent config file formats, at the cost of a bit of manual
copy&paste when the configs are initially created (e.g. copying
database settings from backed-up database.yml)