Do not fail if service is not defined/configured
If a service does not have a local port mapping/does not exist, skip
over it with a warning rather than failing.

This prevents failure if a user removes a service from a compose config;
we do not insist that they use a service in any given enviroment just
because we detect the relevant gem.