Require environment-specified listen port
Do not hardcode the listen port for nginx proxy; require that this is
set by environment as `LISTEN_PORT`. Provide a default `.env` file which
includes `LISTEN_PORT` defaulted to `3000`.

Rename all references to "application" to "app". It seems that "app" is
the standard directory for mounting applications in containers so we
will go with the consensus and be consistent.

Fix a few misc. bugs and simplify the `wait` Makefile logic by checking
the environment in the Makefile and allowing each `wait-foo` command to
decide whether it should run.