Attempt to connect to a network for our services
In the test environment, try to connect to the same network that our
dependencies are running in. This allows running in e.g. Jenkins where
we may already be running in a container (with a mounted
`/var/run/docker.sock`)

Create and use a `docker-compose.local.yml` if it exists (not included
in git) for local overrides, e.g. for local port mappings on services.

Fix tests to match new port strategy.