Fix misc issues with sidecar mode
When sidecar mode is on, store the randomly-generated project name
suffix in `orchestration/.sidecar_suffix`. If this file exists, sidecar
mode will always be assumed to be on and the suffix in the file will be
used.

This allows doing e.g. `make test` and then `make stop env=test`