Fix deployment - use correct compose project name
When running the `deploy` target we want to completely ignore sidecar
mode otherwise the COMPOSE_PROJECT_NAME used to generate our config will
include the sidecar suffix which will then be used to name the network
for that stack. We want the stack name and network name to be the same
(i.e. without a sidecar suffix).