Clear DATABASE_URL from environment before running healthchecks
If any of DATABASE_URL, TEST_DATABASE_URL, or DEVELOPMENT_DATABASE_URL
are set in environment, clear them before running healthchecks (and
restore afterwards) to ensure we do not attempt to connect to an
external database when verifying container availability. Always use the
values that are hardcoded into `config/database.yml` (etc.).