Only skip dumping compose logs if build/push are only command
We want to dump compose logs in cases where build/push are used in
conjunction with other commands, e.g. `make test build push`.

If we only do `make build` or `make push` then compose logs are not
useful so we skip. This still means that `make build push` will output
compose logs but this may be tricky to solve and is not really such a
big problem.