Skip loading environment file for deploy recipe
Environment file for deploy must be in Docker env file format, not
shell format. This means that a valid Docker env file could cause
breakage if sourced into a shell (e.g. if unquoted variables contain
spaces). Since Docker variables should never be quoted (they are always
interpreted literally, including spaces), and since Docker env files are
only used for deployment (local .env file is used for development only
and is sourced into the shell) it is safer to skip the env file
completely - the only variable we need for the deploy recipe is
`RAILS_ENV` so we read that separately into the environment. The
`--env-file` argument is passed to `docker-compose` when rendering the
deployment config, so there is no other benefit to sourcing this into
the environment at deploy time.

Set a placeholder SECRET_KEY_BASE to ensure various Rake commands do not
fail when running in production environment (e.g. at deploy time). Set
default Dockerfile to use the same `placeholder-secret` value for
consistency.